requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
cloudinary==1.29.0
Django==4.0.5
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.1.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
